requirements.txt
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-ldap==3.4.4
